+++ fragmenter/tests/test_fragmenter.py
@@ -4,9 +4,10 @@ Unit and regression test for the fragmenter package.
 
 # Import package, test suite, and other packages as needed
 import fragmenter
+from fragmenter import chemi
+from cmiles import to_canonical_smiles_oe
 import sys
 import unittest
-from openmoltools import openeye
 from fragmenter.tests.utils import get_fn, has_openeye
 
 
@@ -22,9 +23,9 @@ class TestFragment(unittest.TestCase):
         smiles = 'NC(C)(F)C(=O)O'
         isomeric_smiles_r = 'N[C@](C)(F)C(=O)O'
         isomeric_smiles_s = 'N[C@@](C)(F)C(=O)O'
-        mol_1 = openeye.smiles_to_oemol(smiles)
-        mol_2 = openeye.smiles_to_oemol(isomeric_smiles_r)
-        mol_3 = openeye.smiles_to_oemol(isomeric_smiles_s)
+        mol_1 = chemi.smiles_to_oemol(smiles)
+        mol_2 = chemi.smiles_to_oemol(isomeric_smiles_r)
+        mol_3 = chemi.smiles_to_oemol(isomeric_smiles_s)
 
 
         self.assertFalse(fragmenter.fragment._generate_fragments(mol_1))
@@ -35,7 +36,7 @@ class TestFragment(unittest.TestCase):
     def test_expand_protonation_states(self):
         """Test expand protonation states"""
         smiles = 'C5=C(C1=CN=CC=C1)N=C(NC2=C(C=CC(=C2)NC(C3=CC=C(C=C3)CN4CCN(CC4)C)=O)C)N=C5'
-        molecule = openeye.smiles_to_oemol(smiles)
+        molecule = chemi.smiles_to_oemol(smiles)
         protonation = fragmenter.fragment._expand_states(molecule)
         protonation_1 = {'Cc1ccc(cc1Nc2nccc(n2)c3ccc[nH+]c3)NC(=O)c4ccc(cc4)CN5CCN(CC5)C',
                          'Cc1ccc(cc1Nc2nccc(n2)c3ccc[nH+]c3)NC(=O)c4ccc(cc4)CN5CC[NH+](CC5)C',
@@ -55,7 +56,7 @@ class TestFragment(unittest.TestCase):
                          'Cc1ccc(cc1[N-]c2nccc(n2)c3cccnc3)NC(=O)c4ccc(cc4)C[NH+]5CC[NH+](CC5)C'}
         protonation_2 = set()
         for mol in protonation:
-            protonation_2.add(fragmenter.utils.create_mapped_smiles(mol, tagged=False, explicit_hydrogen=False))
+            protonation_2.add(to_canonical_smiles_oe(mol, mapped=False, explicit_hydrogen=False, isomeric=True))
 
         intersection = protonation_1.intersection(protonation_2)
         self.assertEqual(len(intersection), len(protonation_1))
@@ -65,8 +66,8 @@ class TestFragment(unittest.TestCase):
         """Test expand tautomer"""
         smiles_1 ='c1ccc2c(c1)C=CCC2=O'
         smiles_2 = 'c1ccc2c(c1)cccc2O'
-        molecule_1 = fragmenter.utils.smiles_to_oemol(smiles_1)
-        molecule_2 = fragmenter.utils.smiles_to_oemol(smiles_2)
+        molecule_1 = chemi.smiles_to_oemol(smiles_1)
+        molecule_2 = chemi.smiles_to_oemol(smiles_2)
         tautomers_1 = fragmenter.fragment.expand_states(molecule_1, protonation=False, tautomers=True, stereoisomers=False)
         tautomers_2 = fragmenter.fragment.expand_states(molecule_2, protonation=False, tautomers=True, stereoisomers=False)
 
@@ -74,7 +75,7 @@ class TestFragment(unittest.TestCase):
 
     def test_expand_enantiomers(self):
         smiles = 'CN(C)C/C=C/C(=O)NC1=C(C=C2C(=C1)C(=NC=N2)NC3=CC(=C(C=C3)F)Cl)O[C@H]4CCOC4'
-        molecule = openeye.smiles_to_oemol(smiles)
+        molecule = chemi.smiles_to_oemol(smiles)
         stereoisomers = fragmenter.fragment._expand_states(molecule, enumerate='stereoisomers')
 
         stereoisomers_1 = {'CN(C)C/C=C/C(=O)Nc1cc2c(cc1O[C@@H]3CCOC3)ncnc2Nc4ccc(c(c4)Cl)F',
@@ -84,7 +85,7 @@ class TestFragment(unittest.TestCase):
 
         stereoisomers_2 = set()
         for mol in stereoisomers:
-            stereoisomers_2.add(fragmenter.utils.create_mapped_smiles(mol, tagged=False, explicit_hydrogen=False))
+            stereoisomers_2.add(to_canonical_smiles_oe(mol, mapped=False, explicit_hydrogen=False, isomeric=True))
         intersection = stereoisomers_1.intersection(stereoisomers_2)
         self.assertEqual(len(intersection), len(stereoisomers_1))
         self.assertEqual(len(intersection), len(stereoisomers_2))
